refactor: aplicando SRP em criar tipo equipamento, com validator
